+++ AffirmSDK/AffirmAsLowAsButton.m
@@ -5,14 +5,17 @@
 //  Copyright © 2017 Affirm. All rights reserved.
 //
 
+@import SafariServices;
+
 #import "AffirmAsLowAsButton.h"
 #import "AffirmPromoModalViewController.h"
+#import "AffirmConfiguration+Protected.h"
+#import "AffirmUtils.h"
 
 @interface AffirmAsLowAsButton()
 
 @property (nonatomic, strong) NSDecimalNumber *amount;
-@property (nonatomic, strong) NSString *promoID;
-
+@property (nonatomic, assign) BOOL showPrequal;
 @end
 
 @implementation AffirmAsLowAsButton
@@ -38,19 +41,31 @@
                 affirmColor:(AffirmColorType)affirmColor
                 maxFontSize:(CGFloat)maxFontSize
                    callback:(void(^)(BOOL alaEnabled, NSError *error))callback {
-    self.amount = [amount copy];
-    
-    [AffirmAsLowAs getAffirmAsLowAsForAmount:amount promoId:self.promoID affirmLogoType:affirmLogoType affirmColor:affirmColor callback:^(NSString *asLowAsText, UIImage *logo, NSError *error, BOOL success) {
-        [self setAttributedTitle:[AffirmAsLowAs appendLogo:logo toText:asLowAsText font:maxFontSize logoType:affirmLogoType] forState:UIControlStateNormal];
-        [self setAccessibilityLabel:asLowAsText];
+    if (amount && ![[NSDecimalNumber notANumber] isEqualToNumber:amount]) {
+        self.amount = [NSDecimalNumber decimalNumberWithDecimal:[[AffirmNumberUtils decimalDollarsToIntegerCents:amount] decimalValue]];
+    } else {
+        self.amount = nil;
+    }
+
+    [AffirmAsLowAs getAffirmAsLowAsForAmount:self.amount promoId:self.promoID affirmLogoType:affirmLogoType affirmColor:affirmColor callback:^(NSString *ala, UIImage *logo, BOOL showPrequal, NSError *error, BOOL success) {
+        [self setAttributedTitle:[AffirmAsLowAs appendLogo:logo toText:ala font:maxFontSize logoType:affirmLogoType] forState:UIControlStateNormal];
+        [self setAccessibilityLabel:ala];
         [self layoutIfNeeded];
+        self.showPrequal = showPrequal;
         callback(success, error);
     }];
 }
 
 - (void)_showALAModal {
-    AffirmPromoModalViewController *promoModal = [AffirmPromoModalViewController promoModalControllerWithModalId:self.promoID amount:self.amount];
-    [self.presentingViewController presentViewController:promoModal animated:true completion:nil];
+    if (self.showPrequal) {
+        NSString *prequalURL = [AffirmConfiguration sharedConfiguration].affirmPrequalURL;
+        NSString *url = [NSString stringWithFormat:@"%@?public_api_key=%@&unit_price=%@&promo_external_id=%@&isSDK=true&use_promo=True", prequalURL, [AffirmConfiguration sharedConfiguration].publicAPIKey, self.amount, self.promoID];
+        SFSafariViewController *vc = [[SFSafariViewController alloc] initWithURL:[NSURL URLWithString:url]];
+        [self.presentingViewController presentViewController:vc animated:true completion:nil];
+    } else {
+        AffirmPromoModalViewController *promoModal = [AffirmPromoModalViewController promoModalControllerWithModalId:self.promoID amount:self.amount];
+        [self.presentingViewController presentViewController:promoModal animated:true completion:nil];
+    }
 }
 
 @end